Move app Netmap private API views to nav.web.netmap.api module
